requirements.txt
pathlib
pandas
plotly.express
streamlit